+++ next.config.js
@@ -9,7 +9,7 @@ module.exports = {
         source: '/api/:path*',
         destination:
           process.env.NODE_ENV === 'development'
-            ? 'http://127.0.0.1:5000/api/:path*'
+            ? 'http://127.0.0.1:5254/api/:path*'
             : '/api/',
       },
     ]
