+++ package.json
@@ -1,7 +1,9 @@
 {
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "flask-dev": "FLASK_DEBUG=1 pip3 install -r requirements.txt && python3 -m flask --app api/openrocket run -p 5254",
+    "next-dev": "next dev",
+    "dev": "concurrently \"npm run next-dev\" \"npm run flask-dev\"",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -25,6 +27,7 @@
     "autoprefixer": "10.4.16",
     "cloudinary": "^1.41.0",
     "clsx": "^2.1.0",
+    "concurrently": "^8.2.2",
     "date-fns": "2.30.0",
     "highlight.js": "^11.9.0",
     "neo4j-driver": "^5.15.0",
